Alpha beta fix
This should be a correction of your alpha-beta function. To take with a grain of salt as I can't compile it on my machine.

You will have to implement a quiescence search, and your Static Evaluation function must return a score relative to the side to being evaluated (because of the negamax framework).